done delete

+++ web/home.jsp
@@ -93,17 +93,20 @@
                                     <a class="btn btn-success test d-inline"
                                        href="MainServlet?action=ChooseQuiz&QuizID=${id}&type=${requestScope.type}"> Đề ${id}</a>
                                     <c:if test="${sessionScope.user.getRole()=='admin'}">
-                                        <a class="btn btn-primary d-inline" href="MainServlet?action=DeleteQuiz&QuizID=${id}"> Xoá đề </a>
+                                        <a class="btn btn-primary d-inline" href="MainServlet?action=DeleteQuiz&QuizID=${id}" onclick="return confirm('Bạn có chắc không?')"> Xoá đề </a>
 
                                     </c:if>
                                 </div>
 
 
                             </c:forEach>
-                            <div class="mt-3">
-                                <a class="btn btn-success test d-inline"
-                                   href="MainServlet?action=ChooseQuiz&type=${requestScope.type}"> Đề ngẫu nhiên</a>
-                            </div>
+                            <c:if test="${list.size()>0}">
+                                <div class="mt-3">
+                                    <a class="btn btn-success test d-inline"
+                                       href="MainServlet?action=ChooseQuiz&type=${requestScope.type}"> Đề ngẫu nhiên</a>
+                                </div>
+                            </c:if>
+
                         </div>
                     </div>
                 </div>
